[DSV4] Decouple DS V4 Sparse MLA Metadata from DS V3.2 (#44699)
